refactor(app): extract popup-dialog orchestrators (phase 5a)
Move show_warnings_popup, show_lsp_status_popup,
refresh_lsp_status_popup_if_open, build_and_show_lsp_status_popup
(~315 line monolith), show_file_message_popup,
get_text_properties_at_cursor out of buffer_management.rs into a new
app/popup_dialogs.rs sibling. ~444 lines.

These build and dispatch popups as buffer-level ShowPopup events.
None of them are buffer-management proper (open/close/lookup); they
just happened to live there because every popup goes through the
active buffer's popup stack. The biggest one — build_and_show_lsp_status_popup —
is the LSP-server-status popup builder mentioned in the plan as a
prime target for free-function extraction.

buffer_management.rs goes from 3416 -> 2972 lines (-444).
2026 tests still pass.
